+++ server/src/message/Message.h
@@ -54,7 +54,7 @@ namespace message {
         std::unique_ptr<sockaddr> sock_addr;
         std::unique_ptr<EPollEvent *> ePollEvent;
         std::unique_ptr<BlockingQueue<Message> *> blockingQueue;
-        std::unique_ptr<int> id;
+        std::unique_ptr<short> id;
     };
 }
